backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.25
oracledb==2.0.1

pydantic==2.6.4
python-multipart==0.0.9

passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.1
